+++ src/main/java/com/sapo/store_management/repository/CustomerRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
     List<Customer> findByFullnameContainingOrPhoneContaining(String fullname, String phone);
 
-    Optional<Customer> findById(int id);
+    Optional<Customer> findById(Integer id);
 
     public boolean existsByPhone(String phone);
 }
